refactor(server): extract recap session + source-bundle modules, query themes from db
Splits `ProjectRecapJobs.ts` into dedicated `recap-session.ts` (lifecycle / phase
tracking) and `recap-source-bundle.ts` (per-run agent context) modules so the
orchestrator no longer carries hundreds of lines of bundle-construction logic.
Threads the new shapes through `recap-agent-runner.ts` and `ProjectRecap.ts`.

`getRepoContextHandler` now queries `recap_pr_entries` directly via
`inArray(recapId)` instead of relying on `r.entries`, which was always `[]`
(populated by `rowToRecap` with the default empty-entries argument). The agent
now sees the actual theme vocabulary from prior recaps, restoring the
rolling-continuity behavior the field was meant to provide.
